[Chore] update rename & main design, resolves S03P31B304-213

+++ sseukssak/src/components/ModalAddToLibraryDirectory.vue
@@ -1,32 +1,104 @@
 <template>
   <div class="pl-0">
     <v-btn text color="#7288da" v-if="selectedToName" @click="dialog = true">
-      디렉토리 추가 <v-icon>mdi-plus</v-icon>
+      정리 폴더 추가 <v-icon>mdi-plus</v-icon>
     </v-btn>
     <v-dialog
-      width="500px"
+      width="700px"
       v-model="dialog"
       hide-overlay
       transition="dialog-bottom-transition"
       scrollable
     >
       <v-card tile>
-        <v-toolbar flat dark color="primary">
-          <v-btn icon dark @click="closeModal">
+        <v-card-title dark color="#7288da" style="background-color:#7288da; color:white">
+          <v-btn class="mr-3" icon dark @click="closeModal">
             <v-icon>mdi-close</v-icon>
           </v-btn>
-          <v-toolbar-title>디렉토리 추가</v-toolbar-title>
+          정리 폴더 추가
           <v-spacer></v-spacer>
-          <v-toolbar-items>
-            <v-btn dark text @click="createLibrary"> 생성</v-btn>
-          </v-toolbar-items>
-        </v-toolbar>
-
-        <v-card-text style="height: 100%">
+          
+        </v-card-title>
+      
+        <v-card-text class="file-scroller" style="height: 100%">
           <v-container>
             <div style="width: 100%; margin: auto">
-              <div id="type">
-                <v-btn @click="readDir"> 폴더 찾기 </v-btn>{{ directoryDir }}
+              <div>
+                <h2 class="mt-4">1. 정리 폴더 선택</h2>
+                <div class="text-right mb-1">
+                  <v-btn @click="readDir" text color="green"><i class="fas fa-search mr-2"></i>폴더 찾기</v-btn>
+                  <v-dialog class="file-scroller" v-model="dialog3" persistent max-width="500">
+                    <template v-slot:activator="{ on, attrs }">
+                      <v-btn color="green" text v-bind="attrs" v-on="on">
+                        <i class="fas fa-folder-plus mr-2"></i>From 경로에 자동 폴더 생성
+                      </v-btn>
+                    </template>
+                    <v-card>
+                      <v-card-title>
+                        메인 From 경로에 자동 폴더 생성
+                      <!-- <v-tooltip bottom>
+                        <template v-slot:activator="{ on, attrs }">
+                          <i style="color:grey" v-bind="attrs" v-on="on" class="far fa-question-circle fa-2x"></i>
+                        </template>
+                        <span>정리 버튼을 누르면, From 경로에 해당 폴더가 자동 생성됩니다.(같은 폴더명 존재 시, 해당 폴더에 정리됩니다.)</span>
+                      </v-tooltip> -->
+                      </v-card-title>
+                      <v-card-text>
+                        <div>정리 버튼을 누르면, From 경로에 해당 폴더가 자동 생성됩니다.</div>
+                        <div>(같은 폴더명 존재 시, 해당 폴더에 정리됩니다.)</div>
+                      </v-card-text>
+                      <v-card-text>
+                        <v-text-field
+                          v-model="readFromDirName"
+                          label="새 폴더명 입력"
+                        >
+                        </v-text-field
+                      ></v-card-text>
+
+                      <v-card-actions>
+                        <v-spacer></v-spacer>
+                        <v-btn color="green darken-1" text @click="dialog3 = false">
+                          취소
+                        </v-btn>
+                        <v-btn color="green darken-1" text @click="readFromDir">
+                          추가
+                        </v-btn>
+                      </v-card-actions>
+                    </v-card>
+                  </v-dialog>
+                </div>
+                <div class=" ml-3 mb-3 font-weight-bold d-flex align-center">
+                  <img
+                    v-show="directoryDir.length > 0"
+                    src="@/assets/folder-icon.png"
+                    height="30px"
+                    width="30px"
+                    class="mr-2"
+                  />
+                  {{ directoryDir }}
+                </div>
+              </div>
+              <v-divider></v-divider>
+              <h2 class="my-4">2. 정리할 기준 추가</h2>
+              <v-row>
+                <v-col cols="6">
+                  <v-overflow-btn
+                    class="my-2"
+                    :items="filters"
+                    label="정리 기준 선택"
+                    dense
+                    v-model="selectedFilter"
+                  ></v-overflow-btn>
+                </v-col>
+              </v-row>
+              <div id="type" v-show="selectedFilter === '파일 유형'">
+                <div style="width: 80%; margin-left: auto">
+                  <v-text-field v-model="typeAddName" label="확장자 직접 입력(ex. jpg, ppt)" prepend-icon="mdi-pencil" class="pt-0">
+                  <v-icon slot="append" color="green" @click="typeAdd">
+                      mdi-plus
+                  </v-icon>
+                  </v-text-field>
+                </div>
                 <v-select
                   v-model="selectedTypeTags"
                   :items="typeTags"
@@ -34,52 +106,52 @@
                   chips
                   color="red"
                   item-color="red"
-                  label="파일 타입"
+                  label="파일 유형 선택"
                   multiple
                   outlined
                 ></v-select>
-                <div style="width: 50%; margin-left: auto">
-                  <v-text-field v-model="typeAddName" label="확장자 직접입력">
-                    <v-icon slot="append" color="green" @click="typeAdd">
-                      mdi-plus
-                    </v-icon>
-                  </v-text-field>
-                </div>
               </div>
-              <div id="date">
-                <v-select
-                  v-model="selectedDateTags"
-                  :items="dateTags"
-                  deletable-chips
-                  chips
-                  label="날짜"
-                  multiple
-                  outlined
-                >
-                </v-select>
-                <div style="width: 60%; margin-left: auto">
+              <div id="date" v-show="selectedFilter === '날짜'">
+                <div style="width: 80%; margin-left: auto">
                   <v-dialog v-model="dialog2" persistent max-width="290">
                     <template v-slot:activator="{ on, attrs }">
                       <v-text-field
+                        class="pt-0"
                         v-model="dateRangeText"
-                        label="Date range 직접입력"
+                        label="Date range 직접 입력"
                         prepend-icon="mdi-calendar"
                         readonly
                         v-bind="attrs"
                         v-on="on"
                         @click="dates = []"
-                        ><v-icon slot="append" color="green" @click="dateAdd">
-                          mdi-plus
-                        </v-icon></v-text-field
-                      >
+                      ><v-icon slot="append" color="green" @click="dateAdd">
+                        mdi-plus
+                      </v-icon>
+                      </v-text-field>
                     </template>
                     <v-date-picker v-model="dates" range></v-date-picker>
                   </v-dialog>
                 </div>
-                <div style="width: 60%; margin-left: auto"></div>
+              <v-select
+                v-model="selectedDateTags"
+                :items="dateTags"
+                deletable-chips
+                chips
+                label="날짜 분류 선택"
+                multiple
+                outlined
+              >
+              </v-select>
               </div>
-              <div id="name">
-                <v-select
+              <div id="name" v-show="selectedFilter === '파일명'">
+                <div style="width: 100%; margin-left: auto">
+                  <v-text-field v-model="titleAddName" label="파일명 직접 입력" class="pt-0" prepend-icon="mdi-pencil">
+                    <v-icon slot="append" color="green" @click="titleAdd">
+                        mdi-plus
+                    </v-icon>
+                  </v-text-field>
+                </div>
+                <!-- <v-select
                   v-model="selectedTitleTags"
                   :items="titleTags"
                   deletable-chips
@@ -87,24 +159,28 @@
                   label="파일 이름"
                   multiple
                   outlined
-                ></v-select>
-                <div style="width: 60%; margin-left: auto">
-                  <v-text-field v-model="titleAddName" label="파일 이름 입력">
-                    <v-icon slot="append" color="green" @click="titleAdd">
-                      mdi-plus
-                    </v-icon>
-                  </v-text-field>
-                </div>
+                ></v-select> -->
               </div>
+              <v-divider class="mb-10"></v-divider>
               <v-select
+                v-show="selectedTotalTags.length > 0"
                 v-model="selectedTotalTags"
                 :items="totalTags"
                 deletable-chips
                 chips
-                label="선택된 태그들"
+                label="총 정리 기준"
                 multiple
                 outlined
               ></v-select>
+              <div class="text-center">
+                <v-btn
+                  dark
+                  color="#7288da"
+                  @click="createLibrary"
+                >
+                  정리 폴더 추가
+                </v-btn>
+              </div>
             </div>
           </v-container>
         </v-card-text>
@@ -116,6 +192,7 @@
 <script lang='ts'>
 import { Vue, Component, Watch } from "vue-property-decorator";
 import { mapMutations, mapState } from "vuex";
+import Swal from "sweetalert2"
 
 const { dialog } = require("electron").remote;
 
@@ -152,6 +229,7 @@ export default class ModalAddToLibraryDirectory extends Vue {
   directoryDir: string = "";
   dialog: boolean = false;
   dialog2: boolean = false;
+  dialog3: boolean = false;
   typeAddName: string = "";
   titleAddName: string = "";
   typeTags: string[] = [
@@ -169,6 +247,9 @@ export default class ModalAddToLibraryDirectory extends Vue {
   ];
   titleTags: string[] = [];
   selectedTitleTags: string[] = [];
+  filters: string[] = ["파일 유형", "날짜", "파일명"]
+  selectedFilter: string = ""
+  readFromDirName: string = "";
 
   //vuex
   toLibraryList!: ToLibrary[];
@@ -178,17 +259,23 @@ export default class ModalAddToLibraryDirectory extends Vue {
   changeDropToDir!: (dir: string) => void;
   changeToLibraryList!: (newList: ToLibrary[]) => void;
 
-  clickAlert() {
-    alert("click");
-  }
   titleAdd() {
+    if (this.titleAddName === "") {
+      return
+    }
     this.selectedTitleTags.push(this.titleAddName);
     this.titleTags.push(this.titleAddName);
     this.titleAddName = "";
   }
   dateAdd() {
     if (this.dates.length != 2) {
-      alert("날짜를 먼저 선택하거라");
+      Swal.fire({
+        position: "center",
+        icon: "warning",
+        title: "날짜를 선택해주세요",
+        showConfirmButton: false,
+        timer: 1000,
+      });
       return;
     }
     this.selectedDateTags.push(this.dates[0] + "~" + this.dates[1]);
@@ -196,6 +283,9 @@ export default class ModalAddToLibraryDirectory extends Vue {
     this.dates = [];
   }
   typeAdd() {
+    if (this.typeAddName === "") {
+      return
+    }
     let addName = this.typeAddName.toLowerCase();
     if (addName[0] != ".") {
       addName = "." + addName;
@@ -209,7 +299,9 @@ export default class ModalAddToLibraryDirectory extends Vue {
     this.directoryDir = "";
     this.selectedTypeTags = [];
     this.selectedDateTags = [];
-
+    this.readFromDirName = ""
+    this.directoryDir = ""
+    this.selectedFilter = ""
     this.dialog = false;
   }
   readDir() {
@@ -219,37 +311,62 @@ export default class ModalAddToLibraryDirectory extends Vue {
   }
 
   createLibrary() {
-    const tempLibraryList: ToLibrary[] = this.toLibraryList;
+    if (this.directoryDir.trim() === "") {
+      Swal.fire({
+        position: "center",
+        icon: "warning",
+        title: "폴더를 지정해주세요",
+        showConfirmButton: false,
+        timer: 1000,
+      });
+    } else {
+      const tempLibraryList: ToLibrary[] = this.toLibraryList;
 
-    for (let index = 0; index < this.toLibraryNameList.length; index++) {
-      if (tempLibraryList[index].name == this.selectedToName) {
-        tempLibraryList[index].directories.push({
-          path: this.directoryDir,
-          typeTags: this.selectedTypeTags,
-          dateTags: this.selectedDateTags,
-          titleTags: this.selectedTitleTags,
-        });
-        break;
+      for (let index = 0; index < this.toLibraryNameList.length; index++) {
+        if (tempLibraryList[index].name == this.selectedToName) {
+          tempLibraryList[index].directories.push({
+            path: this.directoryDir,
+            typeTags: this.selectedTypeTags,
+            dateTags: this.selectedDateTags,
+            titleTags: this.selectedTitleTags,
+          });
+          break;
+        }
       }
+
+      this.changeToLibraryList(tempLibraryList);
+
+      window.localStorage.setItem(
+        "selectedFromData",
+        JSON.stringify(tempLibraryList)
+      );
+
+      this.libraryDirectories = [];
+      this.directoryDir = "";
+      this.selectedTypeTags = [];
+      this.selectedDateTags = [];
+      this.selectedTitleTags = [];
+      this.readFromDirName = ""
+      this.directoryDir = ""
+      this.selectedFilter = ""
+
+      Swal.fire({
+        position: "center",
+        icon: "success",
+        title: "추가되었습니다",
+        showConfirmButton: false,
+        timer: 1000,
+      });
+
+      this.dialog = false;
     }
-
-    this.changeToLibraryList(tempLibraryList);
-
-    window.localStorage.setItem(
-      "selectedFromData",
-      JSON.stringify(tempLibraryList)
-    );
-
-    this.libraryDirectories = [];
-    this.directoryDir = "";
-    this.selectedTypeTags = [];
-    this.selectedDateTags = [];
-    this.selectedTitleTags = [];
-
-    alert("추가되었습니다.");
-
-    this.dialog = false;
   }
+
+  readFromDir() {
+    this.directoryDir = "%from%" + "\\" + this.readFromDirName;
+    this.dialog3 = false;
+  }
+
   get dateRangeText() {
     return this.dates.join(" ~ ");
   }
